Remove DiagnosticAnalyzerAttributes from abstract classes
This was actually a breaking implementation, that would cause the analysis procedure to attempt to initialize abstract class instances. The abstractions themselves are not removed in the case they appear useful in the future. Until then, they also provide both a shorter name and specify the language they are intended for.